requirements.txt
ccxt==4.3.78
binance-connector==3.6.0
pandas==2.2.2
numpy==1.26.4
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
python-dateutil==2.9.0.post0
pyyaml==6.0.2
